Update draft to correspond to latest YANG module.
o  Redesign of module sub-tree based on review.

o  Modularize some leafs and create typedefs to share with API YANG
   modules.

o  Add module conformance-type, deviation and feature leafs under the
   implementation branch.

o  Change yang-version to be an enum.

o  Add a leaf for module-classification based on
   [I-D.ietf-netmod-yang-model-classification].

o  Normalize enums to be lowercase.

o  Use identities for protocols instead of an enumeration.

o  Make conformance-type optional as not all vendors implement
   [RFC7895].

o  Add a leaf for tree-type based on
   [I-D.ietf-netmod-revised-datastores].

o  Add a reference to contributing to the YANG Catalog at
   yangcatalog.org.

o  Various wording and style changes to the document text.